Various visual fixes
1. Rebake navmesh add missing weapon for Level3
2. Enlarge outer walls to cover map borders
3. Add missing colliders
4. Adjust bullet particle for better visual
5. Move effects scripts to new folder
6. Replace pro builder material with ingame ones
7. Redecorate Level2
