+++ Assets/Source/Game/Scenes/Level3.unity
@@ -10621,6 +10621,68 @@ Mesh:
     offset: 0
     size: 0
     path: 
+--- !u!1001 &1009452014
+PrefabInstance:
+  m_ObjectHideFlags: 0
+  serializedVersion: 2
+  m_Modification:
+    serializedVersion: 3
+    m_TransformParent: {fileID: 1616779572}
+    m_Modifications:
+    - target: {fileID: 7632544431098197865, guid: a3dc26301edc7d24996684ca97c06bc0, type: 3}
+      propertyPath: m_Name
+      value: Bat (1)
+      objectReference: {fileID: 0}
+    - target: {fileID: 8047554303530080302, guid: a3dc26301edc7d24996684ca97c06bc0, type: 3}
+      propertyPath: m_LocalPosition.x
+      value: 55.77
+      objectReference: {fileID: 0}
+    - target: {fileID: 8047554303530080302, guid: a3dc26301edc7d24996684ca97c06bc0, type: 3}
+      propertyPath: m_LocalPosition.y
+      value: 0.19999808
+      objectReference: {fileID: 0}
+    - target: {fileID: 8047554303530080302, guid: a3dc26301edc7d24996684ca97c06bc0, type: 3}
+      propertyPath: m_LocalPosition.z
+      value: 59.48
+      objectReference: {fileID: 0}
+    - target: {fileID: 8047554303530080302, guid: a3dc26301edc7d24996684ca97c06bc0, type: 3}
+      propertyPath: m_LocalRotation.w
+      value: 0.5
+      objectReference: {fileID: 0}
+    - target: {fileID: 8047554303530080302, guid: a3dc26301edc7d24996684ca97c06bc0, type: 3}
+      propertyPath: m_LocalRotation.x
+      value: 0.5
+      objectReference: {fileID: 0}
+    - target: {fileID: 8047554303530080302, guid: a3dc26301edc7d24996684ca97c06bc0, type: 3}
+      propertyPath: m_LocalRotation.y
+      value: 0.5
+      objectReference: {fileID: 0}
+    - target: {fileID: 8047554303530080302, guid: a3dc26301edc7d24996684ca97c06bc0, type: 3}
+      propertyPath: m_LocalRotation.z
+      value: 0.5
+      objectReference: {fileID: 0}
+    - target: {fileID: 8047554303530080302, guid: a3dc26301edc7d24996684ca97c06bc0, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 8047554303530080302, guid: a3dc26301edc7d24996684ca97c06bc0, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.y
+      value: 90
+      objectReference: {fileID: 0}
+    - target: {fileID: 8047554303530080302, guid: a3dc26301edc7d24996684ca97c06bc0, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.z
+      value: 90
+      objectReference: {fileID: 0}
+    m_RemovedComponents: []
+    m_RemovedGameObjects: []
+    m_AddedGameObjects: []
+    m_AddedComponents: []
+  m_SourcePrefab: {fileID: 100100000, guid: a3dc26301edc7d24996684ca97c06bc0, type: 3}
+--- !u!4 &1009452015 stripped
+Transform:
+  m_CorrespondingSourceObject: {fileID: 8047554303530080302, guid: a3dc26301edc7d24996684ca97c06bc0, type: 3}
+  m_PrefabInstance: {fileID: 1009452014}
+  m_PrefabAsset: {fileID: 0}
 --- !u!43 &1030159128
 Mesh:
   m_ObjectHideFlags: 0
@@ -16618,6 +16680,7 @@ Transform:
   - {fileID: 808176073}
   - {fileID: 438335594}
   - {fileID: 859307598}
+  - {fileID: 1009452015}
   - {fileID: 1395926989}
   - {fileID: 1606329335}
   - {fileID: 1674141437}
