--- Assets/Source/Modules/AnimationSystem/PingPong.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: eaa9e0659bf7e024aa9faf5135b510aa
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
