--- Assets/Source/Modules/AnimationSystem/UIMouseFollower.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 8c16df831eb04598bb2837e76653f975
-timeCreated: 1717589908
