+++ Assets/Source/Modules/EffectsSystem/SwingParticleController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3e5e4eb7b5bcfdf409da29c2f848dcf1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
