+++ Assets/Source/Prefabs/Characters/Enemies/Boss/Boss5.prefab
@@ -676,7 +676,7 @@ NavMeshAgent:
   m_Enabled: 1
   m_AgentTypeID: 0
   m_Radius: 0.5
-  m_Speed: 4
+  m_Speed: 5
   m_Acceleration: 200
   avoidancePriority: 50
   m_AngularSpeed: 200
