+++ Assets/Source/Prefabs/LevelDecorations/Maps/Level3 - Stocks fall.prefab
@@ -66079,7 +66079,7 @@ MonoBehaviour:
   m_OverrideVoxelSize: 0
   m_VoxelSize: 0.06666667
   m_MinRegionArea: 2
-  m_NavMeshData: {fileID: 23800000, guid: 904e9cf50f215814a8ccad0a68c7855b, type: 2}
+  m_NavMeshData: {fileID: 23800000, guid: ca2beff0abf5d5546a4f5ce0e7bb5d83, type: 2}
   m_BuildHeightMesh: 0
 --- !u!1 &8863436247706438994
 GameObject:
