+++ Assets/Source/Prefabs/LevelDecorations/Maps/NavMesh-Level16 - Dust off the books.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 12684bd3386edee47a0878216b24c1e8
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 23800000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
