+++ Assets/Source/Prefabs/LevelDecorations/Maps/NavMesh-Level3 - Stocks fall.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ca2beff0abf5d5546a4f5ce0e7bb5d83
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 23800000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
