+++ Assets/Source/Prefabs/LevelDecorations/Maps/NavMesh-Level8 - Bubbling bubbles 1.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e2904c8af49687945abb5e4f416b511a
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 23800000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
