--- Assets/Source/Prefabs/LevelDecorations/Maps/NavMesh-Level8 - Bubbling bubbles.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 2dda95f700d8c9f4bbb45bb95d959122
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 23800000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
