--- Assets/Source/Prefabs/LevelDecorations/NavMesh/NavMesh-Level16 - Dust off the books 1.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 37c65e3201f15ad4a9ed610d3974ba9b
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 23800000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
